move get_dataset_filename. It's fairly fetch-specific
